src/transcriber/requirements.txt
openai==1.69.0
pgvector==0.4.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
ffmpeg-python==0.2.0
git+https://github.com/openai/whisper.git